make assumeRole and createInvalidation their own files
