+++ scripts/deploy/promote-to-production.js
@@ -7,6 +7,8 @@ const AWS = require('aws-sdk')
 const parseArgs = require('./lib/parseArgs')
 const syncBuckets = require('./lib/syncBuckets')
 const { getDeployMetadata } = require('./lib/deploy-metadata')
+const { getAssumeRole } = require('./assume-role')
+const { getCreateInvalidation } = require('./create-invalidation')
 
 const USAGE =
   '\nUsage:\n  node ./scripts/deploy/promote-to-production <project_domain> [--deploy]'
@@ -17,27 +19,10 @@ const dryrun = !flags.includes('--deploy')
 
 assert(projectDomain, USAGE)
 
-const sts = new AWS.STS({ apiVersion: '2011-06-15' })
-
-const productionAssumeRole = () => {
-  return new Promise((resolve, reject) => {
-    sts.assumeRole(
-      {
-        RoleArn: 'arn:aws:iam::043748923082:role/administrator',
-        RoleSessionName: 'promoteToProduction',
-      },
-      (err, data) => {
-        if (err) {
-          reject(err)
-        } else {
-          resolve(data.Credentials)
-        }
-      }
-    )
-  })
-}
-
-productionAssumeRole()
+getAssumeRole(
+  'arn:aws:iam::043748923082:role/administrator',
+  'promoteToProduction'
+)
   .then(credentials => {
     const productionCredentials = new AWS.Credentials({
       accessKeyId: credentials.AccessKeyId,
@@ -70,24 +55,10 @@ productionAssumeRole()
       })
       .then(() => {
         console.log('Promotion to production done\n')
-        const cloudfront = new AWS.CloudFront({
-          apiVersion: '2019-03-26',
-          region: 'us-east-1',
-          credentials: productionCredentials,
-        })
-        const productionCloudfrontArn = `arn:aws:cloudfront::043748923082:distribution/E20OHY6J3BRVIF`
-        const productionDistributionId = productionCloudfrontArn.split('/')[1]
-        const cloudFrontParams = {
-          DistributionId: productionDistributionId,
-          InvalidationBatch: {
-            CallerReference: Date.now().toString(),
-            Paths: {
-              Quantity: 1,
-              Items: ['/*'],
-            },
-          },
-        }
-        return cloudfront.createInvalidation(cloudFrontParams).promise()
+        getCreateInvalidation(
+          productionCredentials,
+          `arn:aws:cloudfront::043748923082:distribution/E20OHY6J3BRVIF`
+        )
       })
       .then(() => {
         console.log('Cache invalidation initiated for production\n')
